# Query result cache to speed up query processing for rarely changing datasets.