--- a/dbms/src/Core/Settings.h
+++ b/dbms/src/Core/Settings.h
@@ -389,6 +389,9 @@ struct Settings : public SettingsCollection<Settings>
     M(SettingBool, enable_scalar_subquery_optimization, true, "If it is set to true, prevent scalar subqueries from (de)serializing large scalar values and possibly avoid running the same subquery more than once.", 0) \
     M(SettingBool, optimize_trivial_count_query, true, "Process trivial 'SELECT count() FROM table' query from metadata.", 0) \
     \
+    M(SettingBool, use_experimental_local_query_cache, false, "Whether to use the cache of local query results.", 0) \
+    M(SettingBool, use_experimental_distributed_query_cache, false, "Whether to use the cache of distributed query results.", 0) \
+    \
     /** Obsolete settings that do nothing but left for compatibility reasons. Remove each one after half a year of obsolescence. */ \
     \
     M(SettingBool, allow_experimental_low_cardinality_type, true, "Obsolete setting, does nothing. Will be removed after 2019-08-13", 0) \