--- a/dbms/src/DataStreams/CacheBlockInputStream.cpp
+++ b/dbms/src/DataStreams/CacheBlockInputStream.cpp
@@ -1,0 +1,1 @@
+#include <DataStreams/CacheBlockInputStream.h>