--- a/dbms/src/Interpreters/ActionsVisitor.cpp
+++ b/dbms/src/Interpreters/ActionsVisitor.cpp
@@ -13,6 +13,7 @@
 #include <DataTypes/DataTypeLowCardinality.h>
 #include <DataTypes/FieldToDataType.h>
 
+#include <DataStreams/CacheBlockInputStream.h>
 #include <DataStreams/LazyBlockInputStream.h>
 
 #include <Columns/ColumnSet.h>
@@ -37,6 +38,7 @@
 #include <Interpreters/convertFieldToType.h>
 #include <Interpreters/interpretSubquery.h>
 #include <Interpreters/DatabaseAndTableWithAlias.h>
+#include <Interpreters/QueryCache.h>
 
 namespace DB
 {
@@ -607,8 +609,33 @@ SetPtr ActionsMatcher::makeSet(const ASTFunction & node, Data & data, bool no_su
         if (!subquery_for_set.source && data.no_storage_or_local)
         {
             auto interpreter = interpretSubquery(right_in_operand, data.context, data.subquery_depth, {});
-            subquery_for_set.source = std::make_shared<LazyBlockInputStream>(
-                interpreter->getSampleBlock(), [interpreter]() mutable { return interpreter->execute().in; });
+            auto create_stream = [&] {
+                return std::make_shared<LazyBlockInputStream>(
+                    interpreter->getSampleBlock(), [interpreter]() mutable { return interpreter->execute().in; });
+            };
+
+            if (data.context.getSettingsRef().use_experimental_local_query_cache)
+            {
+                auto key = QueryCache::getKey(*right_in_operand);
+                auto query_cache = data.context.getQueryCache();
+                auto cache = query_cache->getCache(key, data.context);
+                if (cache)
+                {
+                    subquery_for_set.source = std::make_shared<CacheBlockInputStream>(*cache->blocks);
+                }
+                else
+                {
+                    subquery_for_set.source = create_stream();
+                    auto tables = QueryCache::getRefTables(*right_in_operand, data.context);
+                    if (tables)
+                        subquery_for_set.source->enableQueryCache(key, tables, query_cache);
+                }
+            }
+            else
+            {
+                subquery_for_set.source = create_stream();
+            }
+
 
             /** Why is LazyBlockInputStream used?
               *