--- a/dbms/src/Interpreters/Context.cpp
+++ b/dbms/src/Interpreters/Context.cpp
@@ -26,6 +26,7 @@
 #include <Core/Settings.h>
 #include <Interpreters/ExpressionJIT.h>
 #include <Interpreters/UsersManager.h>
+#include <Interpreters/QueryCache.h>
 #include <Interpreters/Quota.h>
 #include <Dictionaries/Embedded/GeoDictionariesLoader.h>
 #include <Interpreters/EmbeddedDictionaries.h>
@@ -134,6 +135,7 @@ struct ContextShared
     Quotas quotas;                                          /// Known quotas for resource use.
     mutable UncompressedCachePtr uncompressed_cache;        /// The cache of decompressed blocks.
     mutable MarkCachePtr mark_cache;                        /// Cache of marks in compressed files.
+    mutable QueryCachePtr query_cache;                      /// Cache of query results.
     ProcessList process_list;                               /// Executing queries at the moment.
     MergeList merge_list;                                   /// The list of executable merge (for (Replicated)?MergeTree)
     ViewDependencies view_dependencies;                     /// Current dependencies
@@ -1471,6 +1473,28 @@ void Context::dropMarkCache() const
         shared->mark_cache->reset();
 }
 
+void Context::setQueryCache(size_t cache_size_in_bytes, const std::chrono::seconds & expiration_delay)
+{
+    auto lock = getLock();
+
+    if (shared->query_cache)
+        throw Exception("Query cache has been already created.", ErrorCodes::LOGICAL_ERROR);
+
+    shared->query_cache = std::make_shared<QueryCache>(cache_size_in_bytes, expiration_delay);
+}
+
+QueryCachePtr Context::getQueryCache() const
+{
+    auto lock = getLock();
+    return shared->query_cache;
+}
+
+void Context::dropQueryCache() const
+{
+    auto lock = getLock();
+    if (shared->query_cache)
+        shared->query_cache->reset();
+}
 
 void Context::dropCaches() const
 {
@@ -1481,6 +1505,9 @@ void Context::dropCaches() const
 
     if (shared->mark_cache)
         shared->mark_cache->reset();
+
+    if (shared->query_cache)
+        shared->query_cache->reset();
 }
 
 BackgroundProcessingPool & Context::getBackgroundPool()