--- a/dbms/src/Interpreters/DatabaseAndTableWithAlias.h
+++ b/dbms/src/Interpreters/DatabaseAndTableWithAlias.h
@@ -25,6 +25,7 @@ struct DatabaseAndTableWithAlias
     String alias;
 
     DatabaseAndTableWithAlias() = default;
+    DatabaseAndTableWithAlias(String database_, String table_, String alias_);
     DatabaseAndTableWithAlias(const ASTPtr & identifier_node, const String & current_database = "");
     DatabaseAndTableWithAlias(const ASTIdentifier & identifier, const String & current_database = "");
     DatabaseAndTableWithAlias(const ASTTableExpression & table_expression, const String & current_database = "");
@@ -34,6 +35,18 @@ struct DatabaseAndTableWithAlias
 
     /// Check if it satisfies another db_table name. @note opterion is not symmetric.
     bool satisfies(const DatabaseAndTableWithAlias & table, bool table_may_be_an_alias);
+
+    bool operator<(const DatabaseAndTableWithAlias& rhs) const
+    {
+        if (database == rhs.database)
+            if (table == rhs.table)
+                return alias < rhs.alias;
+            else
+                return table < rhs.table;
+        else
+            return database < rhs.database;
+
+    }
 };
 
 using TableWithColumnNames = std::pair<DatabaseAndTableWithAlias, Names>;