--- a/dbms/src/Interpreters/InterpreterDropQuery.h
+++ b/dbms/src/Interpreters/InterpreterDropQuery.h
@@ -34,6 +34,8 @@ private:
 
     BlockIO executeToDictionary(String & database_name, String & table_name, ASTDropQuery::Kind kind, bool if_exists, bool if_temporary, bool no_ddl_lock);
 
+    BlockIO executeToQueryCache();
+
     DatabasePtr tryGetDatabase(String & database_name, bool exists);
 
     DatabaseAndTable tryGetDatabaseAndTable(String & database_name, String & table_name, bool if_exists);