--- a/dbms/src/Parsers/ASTDropQuery.cpp
+++ b/dbms/src/Parsers/ASTDropQuery.cpp
@@ -47,6 +47,8 @@ void ASTDropQuery::formatQueryImpl(const FormatSettings & settings, FormatState 
 
     if (table.empty() && !database.empty())
         settings.ostr << "DATABASE ";
+    else if (query_cache)
+        settings.ostr << "QUERY_CACHE ";
     else if (!is_dictionary)
         settings.ostr << "TABLE ";
     else
@@ -59,7 +61,7 @@ void ASTDropQuery::formatQueryImpl(const FormatSettings & settings, FormatState 
 
     if (table.empty() && !database.empty())
         settings.ostr << backQuoteIfNeed(database);
-    else
+    else if (!query_cache)
         settings.ostr << (!database.empty() ? backQuoteIfNeed(database) + "." : "") << backQuoteIfNeed(table);
 
     formatOnCluster(settings);