… dbms/src/Interpreters/tests/CMakeLists.txt
… dbms/src/Parsers/ParserDropQuery.cpp
… dbms/src/Storages/MergeTree/MergeTreeData.h
… dbms/src/Storages/StorageMergeTree.cpp
… dbms/tests/queries/0_stateless/01035_query_cache.reference
… dbms/tests/queries/0_stateless/01035_query_cache.sql
… dbms/tests/queries/0_stateless/01036_distributed_query_cache.reference
… dbms/tests/queries/0_stateless/01036_distributed_query_cache.sql