I hereby agree to the terms of the CLA available at: https://yandex.ru/legal/cla/?lang=en

For changelog. Remove if this is non-significant change.

Category (leave one):
- New Feature
- Performance Improvement


## Description
There are many unchanging data in analytical databases.
In this case, it is possible to get better performance if ClickHouse stores the calculated data as caches and reuse them.
So I am developing a query cache feature.

**In fact, this is just a basic runnable version**. The reason for open this pull request is that I want to know how the ClickHouse Team views this feature. If it look good to the ClickHouse team, I will work on this feature.

If you want to use this feature, `set use_experimental_query_cache = 1`.


## Performance Test
**Test Data Set**
```SQL
SELECT 
    toYYYYMMDD(time) AS dt, 
    count()
FROM log_table
WHERE (dt >= 20190808) AND (dt <= 20190811)
GROUP BY dt
ORDER BY dt ASC

dt | count()
20190808 │ 22828370
20190809 │ 23090030
20190810 │ 21618814 
20190811 │ 20614539 
```

**First query without cache**
```SQL
SET use_experimental_query_cache = 1;

SELECT 
    uniq(user_id), 
    count(user_id)
FROM log_table
WHERE ((toYYYYMMDD(time) >= 20190810) AND (toYYYYMMDD(time) <= 20191011)) AND (user_id IN (
(
    SELECT user_id
    FROM log_table
    WHERE (toYYYYMMDD(time) >= 20190801) AND (toYYYYMMDD(time) <= 20190811)
    GROUP BY user_id
) AS segment_1))
FORMAT Vertical

Row 1:
──────
uniq(user_id):  853559
count(user_id): 85163641
```
1 rows in set. **Elapsed: 11.325 sec**. Processed 331.20 million rows, 3.97 GB (29.24 million rows/s., 350.19 MB/s.)

**Second query with cache**
```SQL
SET use_experimental_query_cache = 1;

SELECT 
    uniq(user_id), 
    count(user_id)
FROM log_table
WHERE ((toYYYYMMDD(time) >= 20190808) AND (toYYYYMMDD(time) <= 20191009)) AND (user_id IN (
(
    SELECT user_id
    FROM log_table
    WHERE (toYYYYMMDD(time) >= 20190801) AND (toYYYYMMDD(time) <= 20190811)
    GROUP BY user_id
) AS segment_1))
FORMAT Vertical

Row 1:
──────
uniq(user_tuid_hash):  1082853
count(user_tuid_hash): 131082041
```
1 rows in set. **Elapsed: 5.049 sec**. Processed 134.96 million rows, 1.60 GB (26.73 million rows/s., 317.70 MB/s.) 

Notice that time ranges are difference in the two queries.
And second query reusing the result of subquery outperforms the first query.

## Constraints:
It is only applied to the sort of MergeTree engines.


## TODO:
* Update cache if there is any update for data involved to any cache.
* Complete disk option.
* ~~Add TTL for the cache.~~
* Various cache mode such as 'memory', 'disk'
* ~~LRU and max size for the memory~~ and disk for the cache.
* Apply for each session.
* Fine grained locking
* ~~DROP QUERY_CACHE~~