@achimbab 2019-11-13
I added LRUCache to QueryCache.
Even though anyone doesn't reply for this draft I have fixed some codes and reviewed myself.
But, I worry about deadlock or memory leak or any other problems.
I would be very happy if someone would note any comments for this draft.