@achimbab 2019-11-14
I Added `DROP QUERY_CACHE` to delete all query_caches.