@achimbab 2019-11-18
I forcely committed changes to resolve conflict latest master branch and cleaned up commits by rebasing.
* Rebased commits:
  * Implemented basic LRU-Query-Cache.
  * iImplemented 'DROP QUERY_CACHE' to delete all query_caches.